+++ packages/ti/drv/ipc/src/ipc_utils.c
@@ -304,6 +304,7 @@ void IpcUtils_HeapFree(IpcUtils_HeapHandle *pHndl, void* block, uint32_t size)
     return;
 }
 
+
 #ifdef IPC_EXCLUDE_CTRL_TASKS
 /* Fix Me : Move it from here */
 int32_t SystemP_printf(const char* fmt, ...)
@@ -314,7 +315,7 @@ int32_t SystemP_printf(const char* fmt, ...)
 }
 #endif
 
-uint32_t IpcUtils_getMemoryAddress(uint32_t daAddr, uint32_t size)
+uintptr_t IpcUtils_getMemoryAddress(uint32_t daAddr, uint32_t size)
 {
 #if defined(BUILD_MPU1_0) && defined(QNX_OS)
 #include <sys/mman.h>
@@ -324,7 +325,7 @@ uint32_t IpcUtils_getMemoryAddress(uint32_t daAddr, uint32_t size)
     {
         //SystemP_printf("IpcUtils_getMemoryAddress : failed to map..\n");
     }
-    return (uint32_t)(uintptr_t)p;
+    return (uintptr_t)p;
 #else
     return daAddr;
 #endif
